requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
dnspython==2.0.0
finnhub-python==2.2.1
idna==2.10
multidict==4.7.6
pymongo==3.11.0
requests==2.24.0
urllib3==1.25.10
yarl==1.5.1
